feat(package): Adjust package structure
