+++ docs/source/conf.py
@@ -24,7 +24,6 @@ author = 'leesin'
 # The full version, including alpha/beta/rc tags
 release = '0.1.10'
 
-
 # -- General configuration ---------------------------------------------------
 
 # Add any Sphinx extension module names here, as strings. They can be
@@ -47,7 +46,6 @@ language = 'zh_CN'
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = []
-
 
 # -- Options for HTML output -------------------------------------------------
