+++ src/load_m3u8/version.py
@@ -1,2 +1,3 @@
+# _*_coding:utf-8_*_
 __version_info__ = (0, 0, 10)
 __version__ = ".".join(map(str, __version_info__))
