+++ tests/test_generate_m3u8.py
@@ -1,8 +1,10 @@
+# _*_coding:utf-8_*_
+
 import logging
 import sys
 import unittest
 
-from src.generate_m3u8 import resolve
+from load_m3u8.generate import resolve
 
 debug = True
 log_level = logging.DEBUG if debug else logging.INFO
@@ -15,7 +17,7 @@ logging.basicConfig(level=log_level,
 class TestCreateM3U8(unittest.TestCase):
 
     def test_generate(self):
-        video_path = 'c:/tmp/test.ts'
+        video_path = 'D:/test/test.ts'
         create_obj = resolve.CreateM3U8(video_path, m3u8_path=None, hls_time=60, hls_enc_key=None, hls_enc_iv=None,
                                         hls_enc_key_url=None, hls_base_url=None, hls_segment_filename=None)
         create_obj.run()
