+++ tests/test_load_m3u8.py
@@ -3,7 +3,7 @@ import logging
 import sys
 import unittest
 
-from src.load_m3u8 import resolve
+from load_m3u8.load import resolve
 
 debug = True
 log_level = logging.DEBUG if debug else logging.INFO
@@ -17,6 +17,6 @@ class TestLoadM3U8(unittest.TestCase):
 
     def test_download(self):
         m3u8_url = "http://127.0.0.1/test.m3u8"
-        m3u8_url = "C:/tmp/test.m3u8"
-        load_obj = resolve.LoadM3U8(m3u8_url, video_path='C:/tmp/test_m3u8.ts', process_workers=1, thread_workers=1)
+        m3u8_url = "D:/test/test.ts.m3u8"
+        load_obj = resolve.LoadM3U8(m3u8_url, video_path='D:/test/load/test_m3u8.ts', process_workers=1, thread_workers=1)
         load_obj.run()
